+++ charitable/templates/donation-receipt/offline-payment-instructions.php
@@ -18,4 +18,4 @@ if ( ! defined( 'ABSPATH' ) ) {
 /* @var Charitable_Donation */
 $donation = $view_args['donation'];
 
-echo wpautop( do_shortcode( $donation->get_gateway_object()->get_value( 'instructions' ) ) );
+echo wp_kses_post( wpautop( do_shortcode( $donation->get_gateway_object()->get_value( 'instructions' ) ) ) );